finish implementing service

+++ module-service/src/main/java/com/mjc/school/service/NewsService.java
@@ -2,14 +2,14 @@ package com.mjc.school.service;
 
 
 
-import com.mjc.school.repository.NewsRepository;
+import com.mjc.school.dto.NewsDTO;
 
 import java.util.List;
 
 public interface NewsService {
     boolean deleteNewsEntry(long id);
     NewsDTO getNewsById(long id);
-    int createNewsEntry(NewsDTO news);
+    long createNewsEntry(NewsDTO news);
     List<NewsDTO> getAllNews();
     boolean updateNewsEntry(NewsDTO news);
     void init();
